requirements.txt
Django==4.0.1
dj-database-url==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
hugging-face-hub==0.4.0
keras==2.6.0
keras-pickle-wrapper==1.0.5
Keras-Preprocessing==1.1.2
pickle5==0.0.11
Pillow==9.0.0
requests==2.27.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tokenizers==0.11.4
transformers==4.16.1
gunicorn==20.1.0
Markdown==3.3.6
bleach==4.1.0
whitenoise==5.3.0
